+++ CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2025-07-22
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`ssi` - `v2.8.0`](#ssi---v280)
+
+---
+
+#### `ssi` - `v2.8.0`
+
+ - **FEAT**: extend vp universal verifier (#162).
+
+## 2.8.0
+
+ - **FEAT**: extend vp universal verifier (#162).
+
+
+## 2025-07-22
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`ssi` - `v2.7.0`](#ssi---v270)
+
+---
+
+#### `ssi` - `v2.7.0`
+
+ - **FEAT**: extend universal verifiers (#161).
+
+## 2.7.0
+
+ - **FEAT**: extend universal verifiers (#161).
+
+
+## 2025-07-22
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`ssi` - `v2.6.0`](#ssi---v260)
+
+---
+
+#### `ssi` - `v2.6.0`
+
+ - **FEAT**: add customDocumentLoader to vp verifyIntegrity (#160).
+
+## 2.6.0
+
+ - **FEAT**: add customDocumentLoader to vp verifyIntegrity (#160).
+
+
 ## 2025-07-16
 
 ### Changes